+++ src/main/java/es/in2/issuer/backend/backoffice/domain/exception/InvalidStatusException.java
@@ -1,0 +1,7 @@
+package es.in2.issuer.backend.backoffice.domain.exception;
+
+public class InvalidStatusException extends RuntimeException {
+    public InvalidStatusException(String message) {
+        super(message);
+    }
+}