+++ src/main/java/es/in2/issuer/backend/backoffice/domain/model/dtos/RevokeCredentialRequest.java
@@ -1,0 +1,8 @@
+package es.in2.issuer.backend.backoffice.domain.model.dtos;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record RevokeCredentialRequest(
+        @JsonProperty("credentialId") String credentialId,
+        @JsonProperty("listId") int listId) {
+}