+++ src/main/java/es/in2/issuer/backend/shared/domain/util/EndpointsConstants.java
@@ -33,9 +33,11 @@ public class EndpointsConstants {
     // CORS Configuration
     public static final String CORS_OID4VCI_PATH = "/oid4vci/**";
     public static final String CORS_CREDENTIAL_OFFER_PATH = "/oid4vci/v1/credential-offer/**";
+    public static final String STATUS_CREDENTIALS = "/backoffice/v1/credentials/status/**";
 
     // todo: remove these constants if not needed
     public static final String TRUST_FRAMEWORK_ISSUER = "/issuer";
     public static final String DEFERRED_CREDENTIALS = "/api/v1/deferred-credentials";
+
 
 }