+++ src/main/java/es/in2/issuer/backend/shared/infrastructure/repository/CredentialProcedureRepository.java
@@ -1,7 +1,7 @@
 package es.in2.issuer.backend.shared.infrastructure.repository;
 
 import es.in2.issuer.backend.shared.domain.model.entities.CredentialProcedure;
-import es.in2.issuer.backend.shared.domain.model.enums.CredentialStatus;
+import es.in2.issuer.backend.shared.domain.model.enums.CredentialStatusEnum;
 import org.springframework.data.r2dbc.repository.Query;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import org.springframework.stereotype.Repository;
@@ -12,7 +12,7 @@ import java.util.UUID;
 
 @Repository
 public interface CredentialProcedureRepository extends ReactiveCrudRepository<CredentialProcedure, UUID> {
-    Flux<CredentialProcedure> findByCredentialStatusAndOrganizationIdentifier(CredentialStatus credentialStatus, String organizationIdentifier);
+    Flux<CredentialProcedure> findByCredentialStatusAndOrganizationIdentifier(CredentialStatusEnum credentialStatusEnum, String organizationIdentifier);
     @Query("SELECT * FROM issuer.credential_procedure WHERE organization_identifier = :organizationIdentifier ORDER BY updated_at DESC")
     Flux<CredentialProcedure> findAllByOrganizationIdentifier(String organizationIdentifier);
     Mono<CredentialProcedure> findByProcedureIdAndOrganizationIdentifier(UUID procedureId, String organizationIdentifier);