+++ CHANGELOG.md
@@ -4,9 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v2.0.2](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.0.2)
+## [v2.2.0](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.2.0)
 ### Added
 - Add LEARCredentialMachine issuance.
+
+## [v2.1.2](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.1.2)
+### Fixed
+- Fixed url backend in Credential Status
+
+
+## [v2.1.1](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.1.1)
+### Fixed
+- No security for revocation GET endpoint.
+
+## [v2.1.0](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.1.0)
+### Added
+- Add revocation endpoint.
+
+## [v2.0.1](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.0.1)
+### Added
+- Add sign access request.
 
 ## [v2.0.0](https://github.com/in2workspace/in2-issuer-api/releases/tag/v2.0.0)
 ### Added